Fix untappd requiring login and rehash config system
Untappd now requires you be logged in in order to search. No problem,
just add a config option for supplying it with a valid login cookie. Oh
but you can't access the config variables, even if they're global?
Config is now a class through which anything can access the
configuration directives. Previously, this worked by passing them in the
arg[] array, but the Class way is better I think.